[matchup] adds way to get other use in match, and fixes timestamp field.

+++ backend/matchup/models.py
@@ -42,8 +42,8 @@ class Match(models.Model):
     rounds_finished = models.PositiveSmallIntegerField(default=0)
 
     # Timestamps for when the users last sent a message
-    user1_ts = models.PositiveSmallIntegerField(default=0)
-    user2_ts = models.PositiveSmallIntegerField(default=0)
+    user1_ts = models.PositiveIntegerField(default=0)
+    user2_ts = models.PositiveIntegerField(default=0)
 
     def __str__(self):
         return 'match {} with {} users'.format(id, user_count)
